- Add payload define - Add zoom types for IR - Add gimbal set mode - Fix numeric_limits in download_media_files
