+++ CMakeLists.txt
@@ -4,6 +4,18 @@ project (PayloadSDK)
 
 set(CMAKE_BUILD_TYPE Debug)
 
+if (GHADRON)
+set(GHADRON $ENV{GHADRON})
+message("-- GHADRON: ${GHADRON}")
+add_definitions(-DGHADRON=${GHADRON})
+endif(GHADRON)
+
+if (VIO)
+set(VIO $ENV{GHADRON})
+message("-- VIO: ${VIO}")
+add_definitions(-DVIO=${VIO})
+endif(VIO)
+
 find_package(PkgConfig REQUIRED)
 #Use Pkg-config to configure GStreamer
 pkg_search_module(GLIB REQUIRED glib-2.0) 
@@ -32,4 +44,6 @@ add_subdirectory(examples)
 
 
 # Project subdirectory
+if (VIO)
 add_subdirectory(sample_project/demo_project)
+endif(VIO)
