+++ examples/CMakeLists.txt
@@ -1,5 +1,6 @@
 cmake_minimum_required(VERSION 3.5)
 
+if (VIO)
 set(EXAMPLES
     connect.cpp
     load_payload_settings.cpp
@@ -12,12 +13,34 @@ set(EXAMPLES
     get_payload_status.cpp
     eo_set_camera_zoom_focus.cpp
     gimbal_set_mode.cpp
-    ir_zoom.cpp
+    ir_set_camera_zoom.cpp
     ir_capture_image.cpp
     ir_record_video.cpp
     ir_set_palete.cpp
     download_media_files.cpp
 )
+endif(VIO)
+
+if (GHADRON)
+set(EXAMPLES
+    connect.cpp
+    load_payload_settings.cpp
+    # change_payload_settings.cpp
+    eo_capture_image.cpp
+    eo_record_video.cpp
+    # get_video_streaming.cpp
+    move_gimbal_angle_mode.cpp
+    move_gimbal_speed_mode.cpp
+    # get_payload_status.cpp
+    eo_set_camera_zoom_focus.cpp
+    gimbal_set_mode.cpp
+    ir_set_camera_zoom.cpp
+    ir_capture_image.cpp
+    ir_record_video.cpp
+    ir_set_palete.cpp
+    download_media_files.cpp
+)
+endif(GHADRON)
 
 find_package( Threads )
 
@@ -30,6 +53,7 @@ message(STATUS "SUBMODULE_VAR = ${PAYLOADSDK_DIR}")
   
 #include 
 include_directories(${PAYLOADSDK_LIB}
+                    ${PAYLOADSDK_LIB}/payload-define
                     ${PAYLOADSDK_DIR}/third-party/mavlink/include
                     ${PAYLOADSDK_DIR}/gSDK/src/mavlink/include/mavlink/v2.0
                     ${PAYLOADSDK_DIR}/gSDK/src
@@ -45,7 +69,7 @@ foreach(entry ${EXAMPLES})
 
   target_link_libraries(${target}
                      ${PAYLOADSDK_LIB}/libPayloadSDK.a
-                     ${CMAKE_THREAD_LIBS_INIT}
+                    ${CMAKE_THREAD_LIBS_INIT}
                      ${GST_LIBRARIES}
                      ${JSONCPP_LIBRARIES}
                      ${CURL_LIBRARIES}
